requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
sumy==0.11.0
nltk==3.9.1
SQLAlchemy==2.0.32
Werkzeug==3.0.3
